V can now be less than 1 now the simulator does a number of encounter that is a percentage of the number of person that can move multiplied by Vd

+++ src/com/epidemic_simulator/Strategy.java
@@ -8,7 +8,7 @@ import java.util.stream.Collectors;
 public abstract class Strategy implements SimulatorCallBack {
     protected Simulator simulator;
     private ArrayList<Person> population;
-    protected int originalResources;
+    protected long originalResources;
 
     protected HashMap<Integer, HashMap<Person, List<Person>>> encounters;
